Force use of libattr1-dev API
Forwarded: not-needed
Patch-Name: force_xattr_api.diff

Compilation under GNU/kFreeBSD fails because none of upstream's
preprocessor macros matches this combination. It is also not clear
how to properly test for the extended attribute API without
introducing something heavyweight like autoconf. In Debian, however,
the libattr1-dev API is available for all supported architectures,
so it is safe to just hardcode the use of this API. This is what
this patch does.